refactor: WaitingWithRank 패키지 변경
