+++ src/main/java/roomescape/reservation/dto/response/MyReservationResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.vo.WaitingWithRank;
+import roomescape.reservation.dto.WaitingWithRank;
 
 public class MyReservationResponse {
